requirements.txt
fastapi
uvicorn[standard]
openai
faiss-cpu
PyMuPDF
reportlab
numpy
pydantic
